Encapsulate the split view in a new class
Previously,
* BrowserView was responsible for managing the split view.
* ContentsContainer has two web view, and two dev tools, which would
  weird view hierarchy.

This PR introduces `SplitView` class and it encapsulate operations
related to split view. And split view contains two `ContentsContainer`,
so that we don't reuse upstream's contents container as is.

Also this change doesn't intend to change the behavior. The existing
test like `SplitViewBrowserTest` and `SplitViewLayoutManagerTest` would
check the regression.